# docs/swagger.yaml
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.LoginUserPayload:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RegisterUserPayload:
    properties:
      email:
        type: string
      password:
        maxLength: 16
        minLength: 3
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.SuccessResponse:
    properties:
      code:
        type: integer
      status:
        type: string
    type: object
  models.TokenResponse:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Log in a user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: loginUserPayload
        required: true
        schema:
          $ref: '#/definitions/models.LoginUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login user
      tags:
      - User
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      parameters:
      - description: Registration details
        in: body
        name: registerUserPayload
        required: true
        schema:
          $ref: '#/definitions/models.RegisterUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - User
swagger: "2.0"